Fix order of internal message buffer
* In case of a large streamed rpc call the connection broke at ~64kb,
because grpc tried to renew the http stream. In this case there is more
than one message in the internall buffer. The old code pushed new
message to the front of the buffer and also poped them from front. If
there is more than one message in this buffer this will result in a
invalid order.