Dereference the Arc before calling methods on the interface
When calling methods on Arc<MyService> where the method is also defined on Arc (e.g. drop and clone), calling handler_copy.method(o, p) will actually attempt to call the Arc method instead of the method on the service, resulting in a compiler error. This change removes the ambiguity by dereferencing the Arc.